+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add env support for MCPs
 - Add web_search capability
 - Add `o3` model support.
+- Support custom API urls for OpenAI and Anthropic
 
 ## 0.0.3